+++ acceptance/tests/invalid_ssl_config.rb
@@ -1,5 +1,6 @@
 require 'pxp-agent/config_helper.rb'
 require 'pxp-agent/test_helper.rb'
+require 'json'
 
 test_name 'Attempt to start pxp-agent with invalid SSL config'
 
@@ -67,5 +68,48 @@ agents.each_with_index do |agent, i|
       assert_match(/stopped/, result.stdout,
                    "pxp-agent service should stop cleanly when it is running in a loop retrying invalid certs")
     end
+  end
+
+  step 'C97365 - Attempt to run pxp-agent with a different CA to the broker' do
+    step 'Create an alternate ca cert in a temp folder'
+    temp_ca_dir = agent.tmpdir('alt_ca')
+    on agent, puppet("cert list --ssldir #{temp_ca_dir}") # this sets up a CA on the agent
+    windows?(agent) ?
+      alt_ca_cert = "#{temp_ca_dir}\\ca\\ca_crt.pem" :
+      alt_ca_cert = "#{temp_ca_dir}/ca/ca_crt.pem"
+
+    step 'Stop pxp-agent and wipe its existing log file'
+    on agent, puppet('resource service pxp-agent ensure=stopped')
+    on(agent, "rm -rf #{logfile(agent)}")
+
+    step 'Create pxp-agent.conf with an alternate CA cert'
+    pxp_config = pxp_config_hash_using_puppet_certs(master, agent)
+    pxp_config['ssl-ca-cert'] = alt_ca_cert
+    create_remote_file(agent, pxp_agent_config_file(agent), pxp_config.to_json.to_s)
+
+    step 'Start pxp-agent and assert that it does not connect to pcp-broker'
+    on agent, puppet('resource service pxp-agent ensure=running')
+    assert(is_not_associated?(master, "pcp://#{agent}/agent"),
+           "Agent identity pcp://#{agent}/agent for agent host #{agent} should not appear in pcp-broker's inventory " \
+           "when pxp-agent is using the wrong CA cert")
+    expect_file_on_host_to_contain(agent, logfile(agent), 'TLS handshake failed')
+  end
+
+  step 'C97366 - Attempt to connect to pcp-broker without using its certified hostname' do
+    step 'Stop pxp-agent and wipe its existing log file'
+    on agent, puppet('resource service pxp-agent ensure=stopped')
+    on(agent, "rm -rf #{logfile(agent)}")
+
+    step 'Create pxp-agent.conf that connects to pcp-broker using its IP address'
+    pxp_config = pxp_config_hash_using_puppet_certs(master, agent)
+    pxp_config['broker-ws-uri'] = broker_ws_uri(master.ip)
+    create_remote_file(agent, pxp_agent_config_file(agent), pxp_config.to_json.to_s)
+
+    step 'Start pxp-agent and assert that it does not connect to broker'
+    on agent, puppet('resource service pxp-agent ensure=running')
+    assert(is_not_associated?(master, "pcp://#{agent}/agent"),
+           "Agent identity pcp://#{agent}/agent for agent host #{agent} should not appear in pcp-broker's inventory " \
+           "when pxp-agent attempts to connect by broker IP instead of broker certified hostname")
+    expect_file_on_host_to_contain(agent, logfile(agent), 'TLS handshake failed')
   end
 end